prowerki

+++ server/api/visited/delete.delete.ts
@@ -1,4 +1,5 @@
 import prisma from "~/lib/prisma"
+import { authUser } from "~/shared/utils/abilities"
 import { user } from "~/use.vue"
 
 
@@ -6,25 +7,26 @@ interface Body{
     route_id: number
 }
 export default eventHandler(async(event) => {
-    const {route_id} = await readBody<Body>(event)
+    if (authUser) {
+        const {route_id} = await readBody<Body>(event)
 
 
-    const id = await prisma.visited.findFirst({
-        where:{
-            route_id,
-            user_id:user.yandexId
-        },
-        select:{
-            id:true
-        }
-    })
-    if (id) {
-        await prisma.visited.delete({
+        const id = await prisma.visited.findFirst({
             where:{
-                id:+id
+                route_id,
+                user_id:user.yandexId
+            },
+            select:{
+                id:true
             }
-        }) 
+        })
+        if (id) {
+            await prisma.visited.delete({
+                where:{
+                id:+id
+                }
+            }) 
+        }
     }
-   
 
 })
